+++ src/events/index.mjs
@@ -1,3 +1,3 @@
 import './interaction-create.mjs'
 import './ready.mjs'
-import './message-create/index.mjs'
+import './message-create.mjs'
